# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5ea9987e82f82d390e780496
  title: techhubvirtual
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  "/partner/{id}":
    delete:
      x-skaffolder-id: 5ebed5dda940665206b00be0
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5ebeafc1a940665206b00893
      x-skaffolder-resource: Partner
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Partner
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5ebed5dda940665206b00bda
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5ebeafc1a940665206b00893
      x-skaffolder-resource: Partner
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Partner
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partner"
    post:
      x-skaffolder-id: 5ebed5dda940665206b00bdb
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5ebeafc1a940665206b00893
      x-skaffolder-resource: Partner
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Partner
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partner"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Partner"
...
  "/startup/{id}":
    delete:
      x-skaffolder-id: 5ebeafdea940665206b0089c
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5ebeaf96a940665206b00883
      x-skaffolder-resource: Startup
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Startup
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5ebeafdda940665206b00898
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5ebeaf96a940665206b00883
      x-skaffolder-resource: Startup
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Startup
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Startup"
    post:
      x-skaffolder-id: 5ebeafdda940665206b00896
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5ebeaf96a940665206b00883
      x-skaffolder-resource: Startup
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Startup
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Startup"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Startup"
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Partner:
      x-skaffolder-id: 5ebeafc1a940665206b00893
      x-skaffolder-id-db: 5ea9987e82f82d390e780497
      x-skaffolder-id-entity: 5ebeafc1a940665206b00892
      x-skaffolder-url: /partner
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        Description:
          type: string
          x-skaffolder-id-attr: 5ebed5d8a940665206b00bd6
          x-skaffolder-type: String
          x-skaffolder-required: true
        Email:
          type: string
          x-skaffolder-id-attr: 5ebed5d8a940665206b00bd7
          x-skaffolder-type: String
          x-skaffolder-required: true
        Name:
          type: string
          x-skaffolder-id-attr: 5ebed5d8a940665206b00bd5
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - Description
        - Email
        - Name
    Startup:
      x-skaffolder-id: 5ebeaf96a940665206b00883
      x-skaffolder-id-db: 5ea9987e82f82d390e780497
      x-skaffolder-id-entity: 5ebeaf96a940665206b00882
      x-skaffolder-url: /startup
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        cellnumber:
          type: number
          x-skaffolder-id-attr: 5ebeb01da940665206b008a4
          x-skaffolder-type: Number
          x-skaffolder-required: true
        description:
          type: string
          x-skaffolder-id-attr: 5ebeb0e0a940665206b008a8
          x-skaffolder-type: String
          x-skaffolder-required: true
        email:
          type: string
          x-skaffolder-id-attr: 5ebeb01da940665206b008a3
          x-skaffolder-type: String
          x-skaffolder-required: true
        name:
          type: string
          x-skaffolder-id-attr: 5ebeb01da940665206b008a2
          x-skaffolder-type: String
          x-skaffolder-required: true
        website:
          type: string
          x-skaffolder-id-attr: 5ebeb0e0a940665206b008a9
          x-skaffolder-type: String
      x-skaffolder-relations:
        follows:
          x-skaffolder-id: 5ebeb057a940665206b008a7
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 5ebeaf96a940665206b00882
          x-skaffolder-ent2: 5ebeaf96a940665206b00882
      required:
        - _id
        - cellnumber
        - description
        - email
        - name
...